LargeNetworkTest for now also does a Performance test.
